fix: ft_get_next_line_bonus.c
Inside the gnl, the head(the static variable) was been set to NULL at the begin of the code. So every time that the gnl had to read a long line, when it read the first chunk, the variable was reseted at the second chunk, losing content in this process.